requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
dj-static==0.0.6
Django==3.0.2
django-cors-headers==3.2.1
django-rest-authtoken==1.2.4
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
Flask==1.1.1
Flask-Mysql-Html==1.0.2
Flask-MySQLdb==0.2.0
gunicorn==20.0.4
httpie==2.0.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.4
Pillow==7.0.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
PyMySQL==0.9.3
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==0.16.0
wrapt==1.11.2
